add solution dir
